Added reverseStack and enemyMove functions for enemy  maze navigation logic.
Added reverseStack which takes the finishPath stack used when generating the maze to find the finish and reverse it for the enemy to use from start to finish. Added the enemyMove function which implements logic so to get decide what direction the enemy will move to next according to finishPath. Also added logic in main for the visual representation of the enemy moving through the maze to the finish. Might move this to enemyMove function, since they both use the similar logic with the code in main feeling a little inefficient.