Redid the the generateMaze function using a DFS algorithm. Added a simple points system and added lose screen.
Changed the generateMaze function to a DFS algorithm similar to the navigateMaze function. Originally it was hard coded and a fully connected maze took multiple attempts. Also added a simple points system and added a lose screen. Made it so player can hold a direction instead of having to press a key for every cell.